updating examples

+++ testRecordData/main.cpp
@@ -14,15 +14,15 @@
 #include <cstdlib>
 
 // default background frame rate
-const int BG_RATE {10};
-mixr::simulation::Station* station {};
+const int BG_RATE{10};
+mixr::simulation::Station* station{};
 
 // station builder
 mixr::simulation::Station* builder(const std::string& filename)
 {
    // read configuration file
-   int num_errors {};
-   mixr::base::Object* obj {mixr::base::edl_parser(filename, factory, &num_errors)};
+   int num_errors{};
+   mixr::base::Object* obj{mixr::base::edl_parser(filename, factory, &num_errors)};
    if (num_errors > 0) {
       std::cerr << "File: " << filename << ", number of errors: " << num_errors << std::endl;
       std::exit(EXIT_FAILURE);
@@ -61,11 +61,11 @@ void updateDataCB(int msecs)
    glutTimerFunc(msecs, updateDataCB, msecs);
 
    // Current time
-   const double time {mixr::base::getComputerTime()};
+   const double time{mixr::base::getComputerTime()};
 
    // Compute delta time
-   static double time0 {time};   // N-1 Time
-   const double dt {static_cast<double>(time - time0)};
+   static double time0{time};   // N-1 Time
+   const double dt{time - time0};
    time0 = time;
 
    station->updateData(dt);
@@ -93,8 +93,8 @@ int main(int argc, char* argv[])
    station->event(mixr::base::Component::RESET_EVENT);
 
    // set timer for the background tasks
-   const double dt {1.0 / static_cast<double>(BG_RATE)};
-   const int msecs {static_cast<int>(dt * 1000)};
+   const double dt{1.0 / static_cast<double>(BG_RATE)};
+   const int msecs{static_cast<int>(dt * 1000)};
 
    // ensure everything is reset
    station->updateData(dt);
